Added initial logger implementation.
Moved old modules to `troopjs-contrib-browser` (for now). See troopjs/troopjs#76
